FlowShifter progress, made a bare-bones Path class
Refined the FlowShifter class and implemented an abstract
ConvexComboFlowShifter subclass that inherits from the FlowShifter class
and shifts flows according to Eq. 12.20, with  the determination of
lambdas left to the subclass implementation. Also implemented a basic
MSAFlowShifter that inherits from ConvexComboFlowShifter and chooses a
lambda that increases the denominator by 1 each iteration.